% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharma-pukkala-volume.R
\name{forest_regulation_branch_parameters}
\alias{forest_regulation_branch_parameters}
\title{Forest Regulation branch-volume parameters}
\usage{
forest_regulation_branch_parameters()
}
\value{
A data frame containing branch-volume parameters.
}
\description{
Returns the branch-volume ratio parameters prescribed in
Schedule 9 of Nepal's Forest Regulations 2079.
}
