% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-summaries.R
\name{frtc_dbh_summary}
\alias{frtc_dbh_summary}
\title{Summarize FRTC-supported biomass and carbon by plot and DBH class}
\usage{
frtc_dbh_summary(data)
}
\arguments{
\item{data}{Output from [estimate_frtc_biomass()].}
}
\value{
One row per plot and DBH class.
}
\description{
Summarize FRTC-supported biomass and carbon by plot and DBH class
}
