% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-summaries.R
\name{frtc_forest_summary}
\alias{frtc_forest_summary}
\title{Summarize biomass and carbon across inventory plots}
\usage{
frtc_forest_summary(data)
}
\arguments{
\item{data}{Output from [frtc_plot_summary()].}
}
\value{
A one-row data frame containing the mean, standard deviation,
  standard error, minimum, and maximum across plots.
}
\description{
Summarize biomass and carbon across inventory plots
}
