% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-summaries.R
\name{frtc_species_summary}
\alias{frtc_species_summary}
\title{Summarize FRTC-supported biomass and carbon by plot and species}
\usage{
frtc_species_summary(data)
}
\arguments{
\item{data}{Output from [estimate_frtc_biomass()].}
}
\value{
One row per plot and entered species.
}
\description{
Summarize FRTC-supported biomass and carbon by plot and species
}
